+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Breaking Changes
-* Update Openlayers to `^v7.1.0` and ol-mapbox-style to `^9.2.1`.
+* Update Openlayers to `^v7.1.0` and ol-mapbox-style to `^9.2.1`. [Issue #157](https://github.com/dlr-eoc/ukis-frontend-libraries/issues/157).
 
 * **@dlr-eoc/map-ol:**
   - Styles from OpenLayers are no longer included in this component [e3e98c1](https://github.com/dlr-eoc/ukis-frontend-libraries/pull/151/commits/e3e98c1a9a04a23c2109a509e902d14c1b6bdc9d). You have to add them to your application manually see [`@dlr-eoc/map-ol/README.md`](projects/map-ol/README.md).
@@ -13,6 +13,7 @@
  - Add function `hideAllPopups` which adds a css class to the popup element [Issue #120](https://github.com/dlr-eoc/ukis-frontend-libraries/issues/120).
  - The function `layerOnEvent` (for popups) now accepts an optional `feature` as the last parameter.
  - The function `getLayerGroups` is now public to retrieve the LayerGroups and Collections from the map.
+ - Add function `setZIndexForLayerAndGroup` (for internal use). This now moves items in collections so the popup detection works in the new implementation.
 
 ### Bug Fixes
 * **@dlr-eoc/map-ol:**